+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.2.0] 2018-10
+
+### Added
+- Allow adding default response headers to strategies.
+- Expand error handling to include Throwable.
+
 ## [4.1.1] 2018-10
 
 ### Fixed